requirements.txt
openai>=1.68.0
python-dotenv>=1.0.1